Set average report to two decimals

+++ src/app/api/graphql/category/index.ts
@@ -22,7 +22,7 @@ export const typeDefs = `#graphql
     totalCount: Int
 
     totalAmountCents: Int
-    averageAmountCents: Int
+    averageAmountCents: Float
   }
 
   input CategoryCreateInput {
